Complete the refactor around per-type allocators and using allocator aware unique ptrs for owning pointers into custom pools instead of raw pointers.
One might notice that we now enforce lifetime validity such that
owning pointers moved from pass ownership around, so only one
owning pointer ever owns a thing at a time. This will save a large
amount of grief and hassle in months and years to come.

I also added an allocator aware owning span as a stand in for the
future standard static_vector coming in C++ 26. Owning spans are
"good enough" for most use cases for static vectors i.e. basically
whenever you never need to resize them.

What remains yet to do is complete the refactoring such that no
raw C pointer ever owns the memory it points to. That will come
in a subsequent PR, as it is quite involved.