Using the earlier sender-receiver refactorings, refactor process_updates() into process_updates_sender so it can be initiated and progress asynchronously signalling the configured receiver when its root node has completed writing to disc.
This is a major step forward to making the public interface
of trie_db useful to the rest of the Monad codebase, as a fully
async capable API lets other operations be performed concurrently
to trie_db operations.

monad_merge_trie_test produces updates 2.7% faster than in the
niall-senders-receivers2 branch! But the test overall from start
to finish is 1.25% slower. So startup and teardown time seems
to have increased in exchange for faster performance. I've tested
this half a dozen times now, the measurements are correct.

I can't explain that at all - we do more work in this branch,
we call more branches, structure sizes have grown, we even do an
indirect branch down a member variable function pointer. This
branch really should be slower.