Defer new i/o initiated during completions to after completions processing. This prevents stack exhaustion if completions launch more new i/o than was completed.
Previously we filled the submission queue to max before doing
anything about draining completions. Now if the completion queue
is getting close to full, we will process a few completions before
submission. This should prevent dropping of completions which would
be a showstopper.

Previously we assumed that ops in flight was always larger or equal
to submission queue entries. This turns out to not always be true,
it would be appear that io_uring may emit a completion before it has
consumed a submission. Hence now ask io_uring about its submission
queue size when deciding whether its submission queue is full not.

Finally, after discussion yesterday during the triedb meeting,
operation `initiate()` no longer can return failure. It instead invokes
the Receiver with the failure. We still return an enum saying what
happened in case operation initiation needs to know. This has enabled
`erased_connected_operation` to provide a type erased `initiate()`,
which is very useful for plumbing code.