Implement `AsyncReadIoWorkerPool`, a pool of kernel threads each with their own `AsyncIO` instance and therefore own io_uring instance and thus can do their own asynchronous i/o locally. This enables triedb to offload compute and i/o work to the kernel thread pool without inter-thread synchronisation needed if nodes need to be read in from storage.
A serious bug was also fixed in this PR whereby deferred initiations
were not taken into account when deciding if an `AsyncIO` had work
pending or not. This could in certain circumstances cause work to
not get initiated.

I also took a pass at the whole thing with:

- valgrind helgrind
- valgrind drd
- Thread Sanitiser

... and fixed a number of theoretical race conditions, and one actual
race condition.