Fix a race in `AsyncReadIoWorkerPool` whereby if completion was initiated on the worker thread, which then initiated receiver execution on the main thread before the worker thread's execution of the operation had completed, the main thread would tear down the i/o state while the worker thread was still using it.
The solution is to defer telling the main thread that completion has
occurred until after the worker thread has finished with the i/o state.
This take the form of a `timed_delay` operation with a zero timeout,
this turns into a noop in io_uring and simply reinvokes completion
slightly later, but this time outside the worker thread's use of the
i/o state.

Whilst diagnosing and fixing this issue, I also fixed all remaining races
as reported by the thread sanitiser and substantially hardened the
correctness checking of use of Sender-Receiver across kernel threads
by formalising exactly what is allowed when with `MONAD_ASSERT`. This
identified my own racy sloppiness, so will surely be useful into the
future.