Fix bug in compaction when parent has single child during update.
Changed `CompactTNode::cached` to always cache child of an update tnode so that
we avoid the corner case bug in `read_single_child_receiver` with compaction.

More about the bug: we compact nodes in a top down fashion and free compacted
node immediately after comapacting that subtrie, but we won't make its parent
until all other child branches finish up, thus there is a corner case where the
freed compacted node is the only child of its parent and needs to be read and
recreated, however there could be a case when the compacted node still locates
in write buffer and we read in an empty buffer.

Solution here is we set `CompactTNode::cached = true` if its parent's type is
`tnode_type::update`.