Infer min and max history from db_metadata root offsets ring buffer
There was a race condition between the time that min and max version are
updated and when the ring buffer is updated. A read within that window
would be off by one. In this change, the root offsets ring buffer write
pointer is treated as the next version. Min and max versions can be
inferred from that. Switching to a single atomic removes this race
condition as possibility.

Ring buffer changes

  * bugfix for above
  * support discontinuous history with root_offsets.fast_forward()
  * support for upserting multiple times to the same root
  * bump metadata to mnd5