Feat: add call tracer
  * Add call tracing intercept to EvmcHost
  * Add CallFrame RLP
  * Introduce new CALL_FRAME nibble type
  * execute now returns an `ExecutionResult`, which includes both Receipts and CallFrames

CallTraces are stored as vector<CallFrame>, not nested. In order to construct
the proper json format (nested CallFrame), some work needs to be done on the
rpc-server end. TLDR: each CallFrame has a depth, and it is put into the vector
in sequence. (e.g. If the depths are 2,3,3,2, then we know that the 3s are the
child frames of the first 2). See CallTracer::to_json and
CallTracer::to_json_helper for naive implementation to reconstruct the nested
frame.

Known Issues
------------

1. For `CREATE` & `CREATE2` , the `output` field of the `CallFrame` should be
the deployed code. However, `evmc_result` doesn’t contain this information. We
could hackily fix the problem by invoking `get_code` , but I don’t think it’s
necessary.

2.  If `evmc_result` is not `EVMC_SUCCESS`, our current implementation only
returns either `“REVERT”` or `“ERROR”` without giving the reason. We can decode
the reason from the `output` field though (Kai and I discussed this earlier for
`eth_call`).

3. Some fields are empty when they should be omitted, see the reference
implementation below of how to make the result matches exactly with `QuickNode`
RPC.

4. I manually limit the number of `call_frames` to be under 100, so that it
won't blow up `triedb's` node, so in the diff there are 2 huge mismatches
because of this.

5. For some reason, there are ~5 cases where `CALL` is labeled
as `STATICCALL` in our traces. I don't understand this currently as all we need
to check should be `evmc_message.flag` . I don't think it's very important at
the moment.

6. Anvil’s `gas_limit` is different than QuickNode’s `gas_limit` for the top
level. I currently stick to QuickNode, but we can change later.

Co-authored-by: Alexander Lee <alex@monad.xyz>