Split include blocks
This commit separates out includes in blocks according to the following
rough order of headers:
- This project
- Third-party libraries
- Standard library

Doing so means we're less likely to implicitly include a header
somewhere and accidentally rely on something being included
transitively. It's also just a bit easier to read and parse what's being
included in each file.