Make precompile backend flexible
Before this PR, monad's precompile backend was dictated entirely by the
silkpre API. Normally, this wouldn't be an issue. However, silkpre is no
longer maintained, and Dencun support requires us to add a new
precompiled contract type.

Remove the `kSilkpreContracts` (defined in their headers) with a lut of
our own. For now, all of our dispatch functions wrap silkpre, but we can
modify existing functions and add new ones without having to modify and
third party code.

One quirk of this design, in the interim, is that execute functions will
need to allocate output buffers and return those to users. This can be
de-duplicated once we move off of silkpre. For the time being, however,
we need to do this, as the silkpre `run` functions allocate output
buffers.