Cancun: EIP-4788: Add support for beacon block root
The EIP specifies calling the contract at `BEACON_ROOTS_ADDRESS`. The
evmc pseudocode looks like the following:

	evmc_message m{
	    .kind = EVMC_CALL,
	    .gas = 30'000'000,
	    .recipient = BEACON_ROOTS_ADDRESS,
	    .sender = SYSTEM_ADDRESS,
	    .input_data = block.header.timestamp,
	    .input_size = sizeof(block.header.timestamp),
	    .code_address = BEACON_ROOTS_ADDRESS};

The contract it executes maintains `HISTORY_BUFFER_LENGTH` storage slots under
the `BEACON_ROOTS_ADDRESS` account. This contract updates the state, and needs
to be done before executing any transactions in the block.

However, our implementation doesn't invoke the EVM. The EIP notes the
following:

"Clients may decide to omit an explicit EVM call and directly set the storage
values". Doing this is a smaller diff, and the contract code is simple enough,
so we opt for that approach.