[core][cmake] only compile Linux-dependent files on Linux
This is not the ideal way to handle this sort of thing, but in the
interest of time, we're taking a short-cut. This (along with other
changes) will allow the target `monad_core` to compile, but if any of
the functions in these translation units are used, a linker error
will occur. This will allow developments that links to `monad_core` to
proceed on other platforms as long as the final binary does not need
any of the Linux functions, e.g., an isolated test suite binary.

The "right way" to do this would involve figuring the "right interface"
to support (perhaps it is even the Linux one) and implementing that
interface on other platforms. This is possible for features like those
in "cpuset.h" or "statm.h". Every operating system has the support for
what these are doing, but most have different APIs.

There is other functionality here that not every OS supports, e.g.,
allowing the user to directly manipulate large page mappings with mmap(2).
Some systems (Linux, FreeBSD, Windows) allow this, but others do not.

These issues will be solved on an as-needed basis, as they come up in
the development workflow.