Instruction-specific wrappers for `call_runtime`.
This patch abstracts invocations of `call_runtime` in the x86
emitter. The interface of the emitter is extended with a family of
specialised instructions, that, under the hood, invokes
`call_runtime`.

Thus, where we used to write the following in `x86.cpp`:

```c++
...
  case Mul:
     emit.call_runtime(remaining_base_gas, monad::runtime::mul<rev>);
     break;
...
```

We now write:

```c++
...
  case Mul:
     emit.mul<rev>(remaining_base_gas);
     break;
...
```

This patch leaves `call_runtime` public in the interface of the
emitter, because it is currently used in the emitter tests to invoke
two bespoke test functions which are defined outside of the emitter.