[event] add the event recorder system
This commit adds the core of the event API to `libs/core/event`. Although
it is initially going to be used for execution events, the recorder
lives in monad_core because the event system is potentially useful for
a number of tasks, e.g., as a performance tracer. In that case, it
could be used with other lower-layer libraries like `libs/db`.

This only adds the recording infrastructure and test for it. Other parts
of the system will land in subsequent commits.