Cachegrind performance report generator
This patch adds a script which can parse the output of cachegrind and
generate a performance report (in json format). The report contains
the raw output from cachegrind as well as the computed metrics (an
instruction estimate) and some metadata (timestamp and commit
hash). The primary reason to include everything in the report is in
the event we want to retroactively change the metric or compute other
metrics.

The cache simulation details are configurable. They are loaded from a
given json file.

```console
usage: cachegrind.py [-h] [--config FILE] [-o FILE] [-v,--verbose] ...

Cachegrind benchmark runner

positional arguments:
  program        Program to benchmark along with its arguments

options:
  -h, --help     show this help message and exit
  --config FILE  Cache details config
  -o FILE        Save the report to this file (default: /dev/stdout)
  -v,--verbose   Toggle verbose output
```