High-level EVM assembler
This patch is a port of the EVM assembler I prototyped in Scheme. It
is implemented here as an embedded domain-specific language in C++ in
the style of a fluent API.

The primary design goal of this API is to provide an ergonomic and
modular notation for writing EVM programs. The API is dubbed `evm_as`,
and lives in the namespace `monad::vm::utils::evm_as`. This namespace
provides "builder objects" for each EVM revision.

A builder object is essentially a glorified vector/list which is
operated upon by method names corresponding to EVM instructions,
e.g. `push`, `dup1`, `callstatic`, etc. The interesting operations are:

* `push(imm)` automatically expands to a `PUSHn imm` instruction,
  where the `n` is the least number of bytes required to represent
  `imm`.
* `spush(imm)` similar, but allows `imm` to be a signed data type
  (i.e. `int64_t`).
* `jumpdest(label)` labels a jump destination.
* `push(label)` expands to `PUSH label` (which can be consumed by
  `JUMP`/`JUMPI` in the `parser-tool`).
* `jump(label)` and `jumpi(label)` expands to a unconditional jump and
  conditional jump, respectively, to the jumpdest labelled by `label`.

Attempting to emit new instructions in builder objects instanitated at
older revision causes `Invalid` instruction to be injected into the
instruction list. For example, instantiated a builder at frontier and
invoking `push0()` (introduced in revision Shanghai) it will yield an
invalid instruction. However, `push(0)` will do the proper thing and
produce a `PUSH1 0`.

The API also provides a validator and two compilers:

* `evm_as::validate` checks that labels of labelled jumps are defined
  and performs stack underflow and overflow checks. The checks are not
  complete, but sound.
* `evm_as::compile` emits a bytecode representation of the given
  builder object. Labelled jumpdests and labelled jumps are
  automatically desugared.
* `evm_as::mcompile` emits a mnemonic representation. In the future it
  will be able to emit annotated programs too.

Builder objects can be composed via `eb1.compose(eb2)`, without
modifying either `eb1` or `eb2`, yielding a third object. This enables
modular composition of programs, which will be utilised in #388.

* Better provenance: Currently error messages are reported with an
  offset into the instruction list. However, this offset may be
  different in a source program, e.g. if there is insignificant
  newlines, etc. Mapping offsets to source positions will be useful in
  other tools, e.g. parser-tool, monad-compiler-explorer...

* Annotate programs with their stack effects. The Scheme prototype is
  capable of annotating source programs with their stack effects, e.g.
```evm
PUSH1 40
PUSH1 2
ADD

becomes:

PUSH1 40      // [40]
PUSH1 2       // [2 40]
ADD           // [(2 + 40)]
```
which is useful for a human programmer to understand and debug EVM
programs. The building blocks for adding this functionality is nearly
in place. I will add it soon.

* More precise validation. The initial validator is rather simple and
  only meant to provide some quick debugging aid to diagnose some
  common mistakes. If proven to be useful, we can improve its
  precision, e.g. by building the basic block structure and performing
  a control flow analysis and typechecking the program.