[event] add `monad_event_iterator_try_copy`
This function tries to copy the event descriptor at a particular
sequence number, returning true only if the slot is still holding the
descriptor with that sequence number. This is used to replace the earlier
block flow id method using the context area.

As a drive-by fix, we replace __PRETTY_FUNCTION__ with __FUNCTION__
in srcloc.h, as the former is overly verbose and adds little because
we have the source code line