Do not sanitize `baseline_execution.cpp` with GCC/G++.
This commit disables sanitizers on
`third_party/evmone/lib/evmone/baseline_execution.cpp`. The
combination of GCC/G++ w. ASAN and `-O2` enabled causes a
mis-compilation in the baseline execution translation unit, when it
calls `evmc_make_result`. The culprit is that ASAN writes a `\276` to
memory on malloc. This value gets propagated to the result structure,
which causes an assertion failure to trigger in our test fixture.