[mce] Exclusive selection of instrumentation device
This patch makes the chosen instrumentation device of `mce` exclusive in the sense that either the wall clock timer is used to instrument the compiler or the cachegrind instrumentation is used, but never both at the same time. Previously, both instrumentations would be used at the same time, but this does not make sense, especially for wall clock measurements.