[event] move some functions to event_ring.h
This moves a few inline functions from event_iterator.h to event_ring.h,
e.g., monad_event_ring_payload_peek. This was done primarily to benefit
the use of these functions in the Rust. In the Rust event ring module,
these static functions are wrapped in linkable symbols, but end up inlined
in the end, via LTO. It makes sense in Rust for them to be defined on the
event ring and not the iterator, although it makes no difference in C.

Before this change, the payload reading API (and try_copy) was exposed
via the iterator API because:

  - It collected everything needed for reading onto a single cache line,
    including the precomputed masks, in `struct monad_event_iterator`

  - It seemed to "make sense" to keep all the reading-related functions
    together in a single header file

In Rust, this had the unfortunate side effect of requiring the event
descriptor type to carry both the event ring and the iterator's reference
lifetimes explicitly. With this change, the iterator's lifetime parameter
can be removed. The precomputed masks are now kept in the event ring also.