Fix use after free accessing rodb cache nodes
Since the ownership of the child nodes is implicitly implied by owning
the root node, share the node pointer only with requests that have the
same parent node. This will imply that they share root ownership as well
and can share the child pointer.

Add parent node to the key of inflights hash to ensure this property.