[event] add execution exec event definitions
These files define the binary formats of execution event payloads. A few
things to note:

Having standardized versions of C-compatible Ethereum data structures
seems useful for reasons beyond just the execution event payloads, so some
of the files are in the execution `core` subdirectory and not the execution
`event` subdirectory (in later PRs, the `event` subdirectory has a few
other files).

  - base_ctypes.h has the definitions of the basic types like the uint256
    type. This file is different than the others: most of the types in it
    are type aliases to higher level C++ types (e.g., intx::uint256) if
    you are in C++, but if you are in C, they are defined in a way that is
    ABI compatible with intx. This is because all these files are also part
    of the software SDK, and we don't want to assume the user has either
    intx or even C++. The comment that the top of the file explains further.

  - eth_ctypes.h has definitions of basic Ethereum types. Stuff goes in here
    when the authoritative source for its definition is Ethereum, not us

  - exec_event_ctypes.h is the only header file in the `event` subdirectory

How do we deal with ABI versioning?

Since we are writing binary payloads and expecting the reader to understand
them, we need to be sure the structure definitions of the writer and the
readers match.

Version detection is done via comparison of the metadata hash value, a
value that is recorded in both event ring files and a constant in your
compiled library. If they match, the payload contents have the same binary
layout.

For now, it is unspecified how exactly the metadata hash changes or what
the upgrade procedure is. In practice I generate these files via code
generation (these structures are all generated code, and come from TOML
files) but it could be done manually and then the metadata hash would
probably be the git commit hash instead. In the current version, the
metadata hash (which might be called the "binary version hash" when
stuff gets better names before the SDK release) comes from this code
generator.