[monad] add the execution event ring init code
Adds the command line parameter for passing the execution event
ring setup, and maps the event ring into the process address space.
The last part that is missing here, is giving the event ring to the
recorder object. That is commented out, and comes in the next commit.
That is also the code that would gracefully munmap the ring at exit,
so this just lets the OS clean it up at process exit time.

The command line parameter is called `--exec-event-ring` instead
of `--exec-event-path` because you can specify both the file path
and the memory size together in a single configuration string with
this form:

    <file-name>[:<descriptor-shift>:<buf-shift>]

For example:

    /var/lib/hugetlbfs/user/monad/pagesize-2MB/monad-exec-events:21:29

would give 2^21 descriptors (~2 million) and 2^29 = 512 MiB to hold
the event payloads. The sizes are optional, in which case you get the
default values. The above values actually are the default values, and
the reason they were chosen is documented.

A single string is used instead of three different command line switches
because if you were to use other event rings (e.g., for other kinds of
recording) there would then be 3*n options (6, 9, etc.) which is a lot
of command line options.

If `<file-name>` does not contain a `/` character, it is opened
relative to the hugetlbfs directory returned by the new utility function
`monad_event_open_ring_dir_fd` in `event_ring_util.h`. If the user passes
`--exec-event-ring` with no argument at all, it uses the default file
name defined in exec_events_ctypes_metadata.c