[cmake] build libblst.a ourselves
The build system for libblst.a is a custom bash script. Previously we used
ExternalProject_Add to build it, then created an imported target (blst::blst)
to link the compiled build artifact.

When targets using libblst.a are built via the Rust build system (build.rs
calling CMake) this intermittently fails in the CI environment and this failure
is not reproducible anywhere else.

Since this is only two files, just build it ourselves for now. The compile
options and definitions we use are the same as are echoed to terminal when
running `build.sh` manually.