[rlp] check the length of extra_data
The `extra_data` Ethereum header field is a bit different than most RLP
fields in Ethereum: it is variable length but has a 32-byte maximum. We
were not checking this. An Ethereum test called

  BlockchainTests.InvalidBlocks/bcInvalidHeaderTest/DifferentExtraData1025.json

is designed to detect this. I am not sure why it passed before, but this
makes the rlp decode fail with DecodeError::Overflow.

Previously, the over-sized decoding would succeed (perhaps it would fail
a later validation check) and then the execution event recording of
BLOCK_START would memcpy outside the bounds of the 32 byte extra_data
destination buffer.

This had been pointed out in the audit previously, but has surfaced now
(as a real SEGFAULT) because the Ethereum test suite can record and verify
the correctness of execution events.