[staking]: Snapshot set and consensus set have same order
This commit constructs the snapshot set in the same order as the
consensus set. Previously, we popped from the back of the consensus set.
This meant the snapshot set ordering would always be the reverse of the
consensus set.

There are no correctness issues. However, the ordering provides us two
benefits:

1. Better UX: Ben pointed out it's confusing that the sets are ordered
   differently.
2. When the consensus set doesn't change over multiple epochs, we would
   be doing unnecessary DB writes by storing the snapshot in reverse.

The added test fails before this commit and passes after.