[staking]: Throw instead of assert in the contract
All the places in the contract that assert can be triggered during eth
call using state overrides and setting the "from" field to override the
msg.sender. In these cases, the contract will assert and eth call will
halt. Throw instead of assert, as eth call catches
`MonadException`.

The two tests added perform the state overrides that trigger the
throws.