Enhance exec events recording during tests
This commit does a few things:

 - The call to the initialization function somehow got lost in a rebase
   and the recording tests never ran at all; this was fixed

 - If the user passes a file name after the --record-exec-events, it
   will create the event ring file with open(<filename>, O_CREAT | O_EXCL)
   instead of the default, which creates an anonymous memfd_create(2)
   descriptor

 - The CLI also gains a --sleep option, to put the process to sleep for
   a specified number of seconds after the event ring is created but
   before the tests are run

The purpose of the two latter enhancements is to give the user time to run
an eventcap process, to capture all the execution events that are written by
the tests. For example, this was used to capture the output of the all the
EIP-7702 tests, to ensure they're recorded correctly.

Terminal 1:

  test/ethereum_test/monad-ethereum-test --fork prague --gtest_filter=\*7702\* \
      --trace_calls --record-exec-events /tmp/test-ring --sleep 10

Terminal 2:

  cmd/eventcap snap -r /tmp/test-ring -o- | zstd -19 > /tmp/snapshot-7702.zst