[staking]: Return early if start_ptr is bogus
A node can be determined to be in the linked list if it's "prev" pointer
is not empty. To prevent a single iteration, return early if the start
pointer is not in the list.

The impact of this issue is very low, because a user would be providing
bad input and get bad output as a result.  I can't imagine a use case
where this would actually happen, as users would be using the output
from a previous call as the input to the next one. However, it's a
simple enough fix that we should do it for UX purposes.

This issue was reported by spearbit.